+++ modules/planning/tasks/dp_st_speed/st_graph_point.h
@@ -36,21 +36,21 @@ class StGraphPoint {
   const STPoint& point() const;
   const StGraphPoint* pre_point() const;
 
-  double reference_cost() const;
-  double obstacle_cost() const;
-  double total_cost() const;
+  float reference_cost() const;
+  float obstacle_cost() const;
+  float total_cost() const;
 
   void Init(const std::uint32_t index_t, const std::uint32_t index_s,
             const STPoint& st_point);
 
   // given reference speed profile, reach the cost, including position
-  void SetReferenceCost(const double reference_cost);
+  void SetReferenceCost(const float reference_cost);
 
   // given obstacle info, get the cost;
-  void SetObstacleCost(const double obs_cost);
+  void SetObstacleCost(const float obs_cost);
 
   // total cost
-  void SetTotalCost(const double total_cost);
+  void SetTotalCost(const float total_cost);
 
   void SetPrePoint(const StGraphPoint& pre_point);
 
@@ -60,9 +60,9 @@ class StGraphPoint {
   std::uint32_t index_s_ = 0;
   std::uint32_t index_t_ = 0;
 
-  double reference_cost_ = 0.0;
-  double obstacle_cost_ = 0.0;
-  double total_cost_ = std::numeric_limits<double>::infinity();
+  float reference_cost_ = 0.0;
+  float obstacle_cost_ = 0.0;
+  float total_cost_ = std::numeric_limits<float>::infinity();
 };
 
 }  // namespace planning
